compiler: Lift _map_data and _make_sections to LangBB
