+++ devito/passes/iet/parpragma.py
@@ -11,7 +11,7 @@ from devito.passes.iet.engine import iet_pass
 from devito.passes.iet.langbase import LangBB, LangTransformer, DeviceAwareMixin
 from devito.symbolics import INT, ccode
 from devito.tools import as_tuple, prod
-from devito.types import Symbol, NThreadsBase, Wildcard
+from devito.types import Symbol, NThreadsBase
 
 __all__ = ['PragmaSimdTransformer', 'PragmaShmTransformer',
            'PragmaDeviceAwareTransformer', 'PragmaLangBB']
@@ -475,41 +475,6 @@ class PragmaDeviceAwareTransformer(DeviceAwareMixin, PragmaShmTransformer):
 class PragmaLangBB(LangBB):
 
     @classmethod
-    def _make_symbolic_sections_from_imask(cls, f, imask):
-        datasize = cls._map_data(f)
-        if imask is None:
-            imask = [FULL]*len(datasize)
-
-        sections = []
-        for i, j in zip(imask, datasize):
-            if i is FULL:
-                start, size = 0, j
-            else:
-                try:
-                    start, size = i
-                except TypeError:
-                    start, size = i, 1
-            sections.append((start, size))
-
-        # Unroll (or "flatten") the remaining Dimensions not captured by `imask`
-        if len(imask) < len(datasize):
-            try:
-                start, size = sections.pop(-1)
-            except IndexError:
-                start, size = (0, 1)
-            remainder_size = prod(datasize[len(imask):])
-            # The reason we may see a Wildcard is detailed in the `linearize_transfer`
-            # pass, take a look there for more info. Basically, a Wildcard here means
-            # that the symbol `start` is actually a temporary whose value already
-            # represents the unrolled size
-            if not isinstance(start, Wildcard):
-                start *= remainder_size
-            size *= remainder_size
-            sections.append((start, size))
-
-        return sections
-
-    @classmethod
     def _make_sections_from_imask(cls, f, imask):
         sections = cls._make_symbolic_sections_from_imask(f, imask)
 
@@ -517,13 +482,6 @@ class PragmaLangBB(LangBB):
         sections = ''.join(sections)
 
         return sections
-
-    @classmethod
-    def _map_data(cls, f):
-        if f.is_Array:
-            return f.symbolic_shape
-        else:
-            return tuple(f._C_get_field(FULL, d).size for d in f.dimensions)
 
     @classmethod
     def _map_to(cls, f, imask=None, queueid=None):
